refactor: Move toValues from InPredicate.cpp to Filter.h
The function toValues removes duplicated values from the vector and
return them in a std::vector. It was used to build an InPredicate. It
will be needed for building NOT IN filters for Iceberg equality delete
read as well, therefore moving it from velox/functions/prestosql/InPred
icate.cpp to velox/type/Filter.h. This commit also renames it to
deDuplicateValues to make it easier to understand.

Alchemy-item: (ID = 910) Iceberg staging hub commit 6/16 - 02e667971bf64020558159970c240d93737c191a